+++ extras/tests/ContinuousStepperTests/powerOff.cpp
@@ -1,8 +1,8 @@
 #include <ContinuousStepper.h>
 #include <TestFixtures.hpp>
 
-TEST_CASE("ContinuousStepper::powerOff()") {
-  ContinuousStepper stepper;
+TEST_CASE("ContinuousStepper<StepperDriver>::powerOff()") {
+  ContinuousStepper<StepperDriver> stepper;
 
   GIVEN("begin(10, 11) was called") {
     stepper.begin(10, 11);
@@ -65,6 +65,10 @@ TEST_CASE("ContinuousStepper::powerOff()") {
       }
     }
   }
+}
+
+TEST_CASE("ContinuousStepper<FourWireStepper>::powerOff()") {
+  ContinuousStepper<FourWireStepper> stepper;
 
   GIVEN("begin(2, 3, 4, 5) was called") {
     stepper.begin(2, 3, 4, 5);