+++ extras/tests/ContinuousStepperTests/powerOn.cpp
@@ -1,8 +1,8 @@
 #include <ContinuousStepper.h>
 #include <TestFixtures.hpp>
 
-TEST_CASE("ContinuousStepper::powerOn()") {
-  ContinuousStepper stepper;
+TEST_CASE("ContinuousStepper<StepperDriver>::powerOn()") {
+  ContinuousStepper<StepperDriver> stepper;
 
   GIVEN("begin(10, 11) was called") {
     stepper.begin(10, 11);
@@ -109,6 +109,10 @@ TEST_CASE("ContinuousStepper::powerOn()") {
       }
     }
   }
+}
+
+TEST_CASE("ContinuousStepper<FourWireStepper>::powerOn()") {
+  ContinuousStepper<FourWireStepper> stepper;
 
   GIVEN("begin(2, 3, 4, 5) was called") {
     stepper.begin(2, 3, 4, 5);