+++ extras/tests/ContinuousStepperTests/spin.cpp
@@ -1,8 +1,8 @@
 #include <ContinuousStepper.h>
 #include <TestFixtures.hpp>
 
-TEST_CASE("ContinuousStepper::spin()") {
-  ContinuousStepper stepper;
+TEST_CASE("ContinuousStepper<StepperDriver>::spin()") {
+  ContinuousStepper<StepperDriver> stepper;
 
   GIVEN("begin(10, 11) was called") {
     stepper.begin(10, 11);
@@ -108,6 +108,10 @@ TEST_CASE("ContinuousStepper::spin()") {
       }
     }
   }
+}
+
+TEST_CASE("ContinuousStepper<FourWireStepper>::spin()") {
+  ContinuousStepper<FourWireStepper> stepper;
 
   GIVEN("begin(2, 3, 4, 5) was called") {
     stepper.begin(2, 3, 4, 5);