+++ src/ContinuousStepper.h
@@ -1,10 +1,27 @@
 #pragma once
 
-#include "ContinuousStepperBase.h"
+#include "ContinuousStepperImpl.h"
 
 namespace ArduinoContinuousStepper {
-class ContinuousStepper : public ContinuousStepperBase {
+
+class StepperTicker {
 public:
+  StepperTicker(TickListener *listener) : _listener(listener) {}
+
+  void tick() {
+    _listener->tick();
+  }
+
+  void begin() {}
+
+private:
+  TickListener *_listener;
+};
+
+class LoopTicker : public StepperTicker {
+public:
+  LoopTicker(TickListener *listener) : StepperTicker(listener) {}
+
   void loop() {
     if (!_tickPeriod)
       return;
@@ -18,8 +35,7 @@ public:
     }
   }
 
-protected:
-  void setPeriod(unsigned long period) override {
+  void setPeriod(unsigned long period) {
     _tickPeriod = period;
   }
 
@@ -30,6 +46,22 @@ private:
 
   time_t _tickPeriod, _lastTick = 0;
 };
+
+template <class TStepper, class TTicker = LoopTicker>
+class ContinuousStepper : public ContinuousStepperImpl<TStepper, TTicker> {};
+
+template <class TStepper>
+class ContinuousStepper<TStepper, LoopTicker> : public ContinuousStepperImpl<TStepper, LoopTicker> {
+  using TBase = ContinuousStepperImpl<TStepper, LoopTicker>;
+
+public:
+  void loop() {
+    TBase::_ticker.loop();
+  }
+};
+
 } // namespace ArduinoContinuousStepper
 
 using ArduinoContinuousStepper::ContinuousStepper;
+using ArduinoContinuousStepper::FourWireStepper;
+using ArduinoContinuousStepper::StepperDriver;